refactor: merge governance/exceptions into governance/object where it is used
