+++ src/governance/object.h
@@ -6,7 +6,6 @@
 #define BITCOIN_GOVERNANCE_OBJECT_H
 
 #include <governance/common.h>
-#include <governance/exceptions.h>
 #include <governance/vote.h>
 #include <governance/votedb.h>
 #include <sync.h>
@@ -14,6 +13,10 @@
 #include <span.h>
 
 #include <univalue.h>
+
+#include <exception>
+#include <iosfwd>
+#include <string>
 
 class CBLSPublicKey;
 class CDeterministicMNList;
@@ -24,6 +27,65 @@ class CMasternodeMetaMan;
 struct RPCResult;
 
 extern RecursiveMutex cs_main; // NOLINT(readability-redundant-declaration)
+
+enum governance_exception_type_enum_t {
+    /// Default value, normally indicates no exception condition occurred
+    GOVERNANCE_EXCEPTION_NONE = 0,
+    /// Unusual condition requiring no caller action
+    GOVERNANCE_EXCEPTION_WARNING = 1,
+    /// Requested operation cannot be performed
+    GOVERNANCE_EXCEPTION_PERMANENT_ERROR = 2,
+    /// Requested operation not currently possible, may resubmit later
+    GOVERNANCE_EXCEPTION_TEMPORARY_ERROR = 3,
+    /// Unexpected error (ie. should not happen unless there is a bug in the code)
+    GOVERNANCE_EXCEPTION_INTERNAL_ERROR = 4
+};
+
+std::ostream& operator<<(std::ostream& os, governance_exception_type_enum_t eType);
+
+/**
+ * A class which encapsulates information about a governance exception condition
+ *
+ * Derives from std::exception so is suitable for throwing
+ * (ie. will be caught by a std::exception handler) but may also be used as a
+ * normal object.
+ */
+class CGovernanceException : public std::exception
+{
+private:
+    std::string strMessage;
+
+    governance_exception_type_enum_t eType;
+
+    int nNodePenalty;
+
+public:
+    explicit CGovernanceException(const std::string& strMessageIn = "",
+        governance_exception_type_enum_t eTypeIn = GOVERNANCE_EXCEPTION_NONE,
+        int nNodePenaltyIn = 0);
+
+    ~CGovernanceException() noexcept override = default;
+
+    const char* what() const noexcept override
+    {
+        return strMessage.c_str();
+    }
+
+    const std::string& GetMessage() const
+    {
+        return strMessage;
+    }
+
+    governance_exception_type_enum_t GetType() const
+    {
+        return eType;
+    }
+
+    int GetNodePenalty() const
+    {
+        return nNodePenalty;
+    }
+};
 
 static constexpr double GOVERNANCE_FILTER_FP_RATE = 0.001;
 static constexpr CAmount GOVERNANCE_PROPOSAL_FEE_TX = (1 * COIN);
